Refactor LLM client for lazy loading
