+++ giskard/llm/__init__.py
@@ -1,4 +1,5 @@
+from .client import set_default_client
 from .config import set_openai_key, set_openai_organization
 from .errors import LLMImportError
 
-__all__ = ["LLMImportError", "set_openai_key", "set_openai_organization"]
+__all__ = ["LLMImportError", "set_openai_key", "set_openai_organization", "set_default_client"]
