+++ giskard/llm/client/base.py
@@ -1,0 +1,29 @@
+from abc import ABC, abstractmethod
+from dataclasses import dataclass
+from typing import Any, Dict, Optional
+
+
+class LLMClient(ABC):
+    @abstractmethod
+    def complete(
+        self,
+        messages,
+        functions=None,
+        model="gpt-4",
+        temperature=0.5,
+        max_tokens=None,
+        function_call: Optional[Dict] = None,
+    ):
+        ...
+
+
+@dataclass
+class LLMOutput:
+    message: Optional[str] = None
+    function_call: Dict[str, str] = None
+
+
+@dataclass
+class LLMFunctionCall:
+    function: str
+    args: Any
