+++ giskard/llm/evaluators/base.py
@@ -3,8 +3,7 @@ from typing import Sequence
 
 from ...datasets.base import Dataset
 from ...models.base.model import BaseModel
-from ..client import LLMClient
-from ..client import llm_client as default_llm_client
+from ..client import LLMClient, get_default_client
 from ..errors import LLMGenerationError
 
 EVALUATE_MODEL_FUNCTIONS = [
@@ -59,7 +58,7 @@ class LLMBasedEvaluator:
         self.eval_prompt = eval_prompt or self._default_eval_prompt
         self.llm_model = llm_model
         self.llm_temperature = llm_temperature
-        self.llm_client = llm_client if llm_client is not None else default_llm_client
+        self.llm_client = llm_client if llm_client is not None else get_default_client()
 
     def _make_evaluate_prompt(self, model: BaseModel, input_vars, model_output):
         return self.eval_prompt.format(
